@@ -1,5 +1,4 @@
-//go:build !race
-// +build !race
+//go:build !race && !windows
 
 package node_exporter //nolint:golint
 
