@@ -8,12 +8,8 @@ package main
 
 import (
 	"github.com/grafana/agent/cmd/internal/flowmode"
-
-	// Adds version information
-	_ "github.com/grafana/agent/pkg/build"
-
+	"github.com/grafana/agent/pkg/build"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/version"
 
 	// Register Prometheus SD components
 	_ "github.com/grafana/loki/clients/pkg/promtail/discovery/consulagent"
@@ -24,10 +20,9 @@ import (
 )
 
 func init() {
-	prometheus.MustRegister(version.NewCollector("agent"))
+	prometheus.MustRegister(build.NewCollector("agent"))
 }
 
 func main() {
-	// TODO: Allow Flow to run as a Windows service.
 	flowmode.Run()
 }
