@@ -12,13 +12,11 @@ import (
 	"strings"
 	"syscall"
 
-	"gopkg.in/yaml.v2"
-
+	"github.com/grafana/agent/pkg/build"
 	"github.com/grafana/agent/pkg/config"
 	"github.com/grafana/agent/pkg/logs"
 	"github.com/olekukonko/tablewriter"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/version"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -45,16 +43,13 @@ import (
 	kclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	kconfig "sigs.k8s.io/controller-runtime/pkg/client/config"
-
-	// Adds version information
-	_ "github.com/grafana/agent/pkg/build"
 )
 
 func main() {
 	cmd := &cobra.Command{
 		Use:     "agentctl",
 		Short:   "Tools for interacting with the Grafana Agent",
-		Version: version.Print("agentctl"),
+		Version: build.Print("agentctl"),
 	}
 	cmd.SetVersionTemplate("{{ .Version }}\n")
 
@@ -65,7 +60,6 @@ func main() {
 		targetStatsCmd(),
 		samplesCmd(),
 		operatorDetachCmd(),
-		templateDryRunCmd(),
 		testLogs(),
 	)
 
@@ -82,7 +76,7 @@ func configSyncCmd() *cobra.Command {
 		Use:   "config-sync [directory]",
 		Short: "Sync config files from a directory to an Agent's config management API",
 		Long: `config-sync loads all files ending with .yml or .yaml from the specified
-directory and uploads them the the config management API. The name of the config
+directory and uploads them through the config management API. The name of the config
 uploaded will be the base name of the file (e.g., the name of the file without
 its extension).
 
@@ -429,40 +423,6 @@ func filterAgentOwners(refs []meta_v1.OwnerReference) (filtered []meta_v1.OwnerR
 		filtered = append(filtered, ref)
 	}
 	return
-}
-
-func templateDryRunCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "template-parse [directory]",
-		Short: "dry run dynamic configuration",
-		Long:  `This will load the dynamic configuration, load configs, run templates and then output the full config as yaml`,
-		Args:  cobra.ExactArgs(1),
-
-		RunE: func(_ *cobra.Command, args []string) error {
-			cmf, err := config.NewDynamicLoader()
-			if err != nil {
-				return err
-			}
-			c := &config.Config{}
-			err = cmf.LoadConfigByPath(args[0])
-			if err != nil {
-				return err
-			}
-			err = cmf.ProcessConfigs(c)
-			if err != nil {
-				return fmt.Errorf("error processing config templates %s", err)
-			}
-
-			outBytes, err := yaml.Marshal(c)
-			if err != nil {
-				return err
-			}
-			fmt.Println(string(outBytes))
-			return nil
-		},
-	}
-
-	return cmd
 }
 
 func testLogs() *cobra.Command {
