@@ -42,8 +42,8 @@ type Arguments struct {
 	// Type indicates how to detect changes to the file.
 	Type Detector `river:"detector,attr,optional"`
 	// PollFrequency determines the frequency to check for changes when Type is
-	// UpdateTypePoll.
-	PollFrequency time.Duration `river:"poll_freqency,attr,optional"`
+	// Poll.
+	PollFrequency time.Duration `river:"poll_frequency,attr,optional"`
 	// IsSecret marks the file as holding a secret value which should not be
 	// displayed to the user.
 	IsSecret bool `river:"is_secret,attr,optional"`
@@ -195,7 +195,7 @@ func (c *Component) Update(args component.Arguments) error {
 	newArgs := args.(Arguments)
 
 	if newArgs.PollFrequency <= 0 {
-		return fmt.Errorf("poll_freqency must be greater than 0")
+		return fmt.Errorf("poll_frequency must be greater than 0")
 	}
 
 	c.mut.Lock()
@@ -249,10 +249,10 @@ func (c *Component) configureDetector() error {
 		})
 	case DetectorFSNotify:
 		c.detector, err = newFSNotify(fsNotifyOptions{
-			Logger:       c.opts.Logger,
-			Filename:     c.args.Filename,
-			ReloadFile:   reloadFile,
-			PollFreqency: c.args.PollFrequency,
+			Logger:        c.opts.Logger,
+			Filename:      c.args.Filename,
+			ReloadFile:    reloadFile,
+			PollFrequency: c.args.PollFrequency,
 		})
 	}
 
