@@ -10,11 +10,13 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/common/model"
 )
 
-// MaxPartialLinesSize is the max buffer size to hold partial lines when
-// parsing the CRI stage format.
-const MaxPartialLinesSize = 100
+const (
+	RFC3339Nano         = "RFC3339Nano"
+	MaxPartialLinesSize = 100 // MaxPartialLinesSize is the max buffer size to hold partial lines when parsing the CRI stage format.lines.
+)
 
 // DockerConfig is an empty struct that is used to enable a pre-defined
 // pipeline for decoding entries that are using the Docker logs format.
@@ -45,7 +47,7 @@ func NewDocker(logger log.Logger, registerer prometheus.Registerer) (Stage, erro
 		{
 			TimestampConfig: &TimestampConfig{
 				Source: "timestamp",
-				Format: "RFC3339Nano",
+				Format: RFC3339Nano,
 			},
 		},
 		{
@@ -59,7 +61,7 @@ func NewDocker(logger log.Logger, registerer prometheus.Registerer) (Stage, erro
 
 type cri struct {
 	// bounded buffer for CRI-O Partial logs lines (identified with tag `P` till we reach first `F`)
-	partialLines    []string
+	partialLines    map[model.Fingerprint]Entry
 	maxPartialLines int
 	base            *Pipeline
 }
@@ -69,30 +71,48 @@ func (c *cri) Name() string {
 	return "cri"
 }
 
-// Run implements the Stage interface.
+// Run implements Stage interface
 func (c *cri) Run(entry chan Entry) chan Entry {
 	entry = c.base.Run(entry)
 
-	in := RunWithSkip(entry, func(e Entry) (Entry, bool) {
+	in := RunWithSkipOrSendMany(entry, func(e Entry) ([]Entry, bool) {
+		fingerprint := e.Labels.Fingerprint()
+
+		// We received partial-line (tag: "P")
 		if e.Extracted["flags"] == "P" {
-			if len(c.partialLines) >= c.maxPartialLines {
+			if len(c.partialLines) > c.maxPartialLines {
 				// Merge existing partialLines
-				newPartialLine := e.Line
-				e.Line = strings.Join(c.partialLines, "\n")
+				entries := make([]Entry, 0, len(c.partialLines))
+				for _, v := range c.partialLines {
+					entries = append(entries, v)
+				}
+
 				level.Warn(c.base.logger).Log("msg", "cri stage: partial lines upperbound exceeded. merging it to single line", "threshold", MaxPartialLinesSize)
-				c.partialLines = c.partialLines[:0]
-				c.partialLines = append(c.partialLines, newPartialLine)
-				return e, false
+
+				c.partialLines = make(map[model.Fingerprint]Entry)
+				c.partialLines[fingerprint] = e
+
+				return entries, false
 			}
-			c.partialLines = append(c.partialLines, e.Line)
-			return e, true
+
+			prev, ok := c.partialLines[fingerprint]
+			if ok {
+				e.Line = strings.Join([]string{prev.Line, e.Line}, "")
+			}
+			c.partialLines[fingerprint] = e
+
+			return []Entry{e}, true // it's a partial-line so skip it.
 		}
-		if len(c.partialLines) > 0 {
-			c.partialLines = append(c.partialLines, e.Line)
-			e.Line = strings.Join(c.partialLines, "\n")
-			c.partialLines = c.partialLines[:0]
+
+		// Now we got full-line (tag: "F").
+		// 1. If any old partialLines matches with this full-line stream, merge it
+		// 2. Else just return the full line.
+		prev, ok := c.partialLines[fingerprint]
+		if ok {
+			e.Line = strings.Join([]string{prev.Line, e.Line}, "")
+			delete(c.partialLines, fingerprint)
 		}
-		return e, false
+		return []Entry{e}, false
 	})
 
 	return in
@@ -115,7 +135,7 @@ func NewCRI(logger log.Logger, registerer prometheus.Registerer) (Stage, error) 
 		{
 			TimestampConfig: &TimestampConfig{
 				Source: "time",
-				Format: "RFC3339Nano",
+				Format: RFC3339Nano,
 			},
 		},
 		{
@@ -139,6 +159,6 @@ func NewCRI(logger log.Logger, registerer prometheus.Registerer) (Stage, error) 
 		maxPartialLines: MaxPartialLinesSize,
 		base:            p,
 	}
-	c.partialLines = make([]string, 0, c.maxPartialLines)
+	c.partialLines = make(map[model.Fingerprint]Entry)
 	return &c, nil
 }
