@@ -11,6 +11,8 @@ import (
 	"github.com/grafana/agent/component/common/loki"
 	flow_relabel "github.com/grafana/agent/component/common/relabel"
 	ht "github.com/grafana/agent/component/loki/source/heroku/internal/herokutarget"
+	"github.com/grafana/agent/pkg/util"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/relabel"
 	sv "github.com/weaveworks/common/server"
@@ -64,8 +66,9 @@ func (lc *ListenerConfig) UnmarshalRiver(f func(interface{}) error) error {
 
 // Component implements the loki.source.heroku component.
 type Component struct {
-	opts    component.Options
-	metrics *ht.Metrics
+	opts          component.Options
+	metrics       *ht.Metrics              // Metrics about Heroku entries.
+	serverMetrics *util.UncheckedCollector // Metircs about the HTTP server managed by the component.
 
 	mut    sync.RWMutex
 	args   Arguments
@@ -78,14 +81,17 @@ type Component struct {
 // New creates a new loki.source.heroku component.
 func New(o component.Options, args Arguments) (*Component, error) {
 	c := &Component{
-		opts:    o,
-		metrics: ht.NewMetrics(o.Registerer),
-		mut:     sync.RWMutex{},
-		args:    Arguments{},
-		fanout:  args.ForwardTo,
-		target:  nil,
-		handler: make(loki.LogsReceiver),
+		opts:          o,
+		metrics:       ht.NewMetrics(o.Registerer),
+		mut:           sync.RWMutex{},
+		args:          Arguments{},
+		fanout:        args.ForwardTo,
+		target:        nil,
+		handler:       make(loki.LogsReceiver),
+		serverMetrics: util.NewUncheckedCollector(nil),
 	}
+
+	o.Registerer.MustRegister(c.serverMetrics)
 
 	// Call to Update() to start readers and set receivers once at the start.
 	if err := c.Update(args); err != nil {
@@ -145,8 +151,15 @@ func (c *Component) Update(args component.Arguments) error {
 			}
 		}
 
+		// [ht.NewHerokuTarget] registers new metrics every time it is called. To
+		// avoid issues with re-registering metrics with the same name, we create a
+		// new registry for the target every time we create one, and pass it to an
+		// unchecked collector to bypass uniqueness checking.
+		registry := prometheus.NewRegistry()
+		c.serverMetrics.SetCollector(registry)
+
 		entryHandler := loki.NewEntryHandler(c.handler, func() {})
-		t, err := ht.NewHerokuTarget(c.metrics, c.opts.Logger, entryHandler, rcs, newArgs.Convert(), c.opts.Registerer)
+		t, err := ht.NewHerokuTarget(c.metrics, c.opts.Logger, entryHandler, rcs, newArgs.Convert(), registry)
 		if err != nil {
 			level.Error(c.opts.Logger).Log("msg", "failed to create heroku listener with provided config", "err", err)
 			return err
