@@ -8,8 +8,9 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 
+	"context"
+
 	"github.com/aws/aws-sdk-go-v2/service/s3"
-	"golang.org/x/net/context"
 )
 
 type watcher struct {
