@@ -2,7 +2,6 @@ package config
 
 import (
 	"bytes"
-	"errors"
 	"flag"
 	"fmt"
 	"os"
@@ -13,6 +12,7 @@ import (
 	"github.com/drone/envsubst/v2"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
+	"github.com/grafana/agent/pkg/build"
 	"github.com/grafana/agent/pkg/config/features"
 	"github.com/grafana/agent/pkg/config/instrumentation"
 	"github.com/grafana/agent/pkg/logs"
@@ -21,7 +21,6 @@ import (
 	"github.com/grafana/agent/pkg/traces"
 	"github.com/grafana/agent/pkg/util"
 	"github.com/prometheus/common/config"
-	"github.com/prometheus/common/version"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 )
@@ -29,14 +28,12 @@ import (
 var (
 	featRemoteConfigs    = features.Feature("remote-configs")
 	featIntegrationsNext = features.Feature("integrations-next")
-	featDynamicConfig    = features.Feature("dynamic-config")
 	featExtraMetrics     = features.Feature("extra-scrape-metrics")
 	featAgentManagement  = features.Feature("agent-management")
 
 	allFeatures = []features.Feature{
 		featRemoteConfigs,
 		featIntegrationsNext,
-		featDynamicConfig,
 		featExtraMetrics,
 		featAgentManagement,
 	}
@@ -343,28 +340,6 @@ func LoadRemote(url string, expandEnvVars bool, c *Config) error {
 	return LoadBytes(bb, expandEnvVars, c)
 }
 
-// LoadDynamicConfiguration is used to load configuration from a variety of sources using
-// dynamic loader, this is a templated approach
-func LoadDynamicConfiguration(url string, expandvar bool, c *Config) error {
-	if expandvar {
-		return errors.New("expand var is not supported when using dynamic configuration, use gomplate env instead")
-	}
-	cmf, err := NewDynamicLoader()
-	if err != nil {
-		return err
-	}
-	err = cmf.LoadConfigByPath(url)
-	if err != nil {
-		return err
-	}
-
-	err = cmf.ProcessConfigs(c)
-	if err != nil {
-		return fmt.Errorf("error processing config templates %w", err)
-	}
-	return nil
-}
-
 func performEnvVarExpansion(buf []byte, expandEnvVars bool) ([]byte, error) {
 	// (Optionally) expand with environment variables
 	if expandEnvVars {
@@ -422,17 +397,6 @@ func Load(fs *flag.FlagSet, args []string, log *server.Logger) (*Config, error) 
 				return loadFromAgentManagementAPI(path, expandArgs, c, log, fs)
 			}
 			return LoadFile(path, expandArgs, c)
-		case fileTypeDynamic:
-			if !features.Enabled(fs, featDynamicConfig) {
-				return fmt.Errorf("feature %q must be enabled to use file type %s", featDynamicConfig, fileTypeDynamic)
-			} else if !features.Enabled(fs, featIntegrationsNext) {
-				return fmt.Errorf("feature %q must be enabled to use file type %s", featIntegrationsNext, fileTypeDynamic)
-			} else if features.Enabled(fs, featRemoteConfigs) {
-				return fmt.Errorf("feature %q can not be enabled with file type %s", featRemoteConfigs, fileTypeDynamic)
-			} else if expandArgs {
-				return fmt.Errorf("-config.expand-env can not be used with file type %s", fileTypeDynamic)
-			}
-			return LoadDynamicConfiguration(path, expandArgs, c)
 		default:
 			return fmt.Errorf("unknown file type %q. accepted values: %s", fileType, strings.Join(fileTypes, ", "))
 		}
@@ -493,7 +457,7 @@ func load(fs *flag.FlagSet, args []string, loader loaderFunc) (*Config, error) {
 	}
 
 	if printVersion {
-		fmt.Println(version.Print("agent"))
+		fmt.Println(build.Print("agent"))
 		os.Exit(0)
 	}
 
