@@ -448,12 +448,6 @@ metrics:
 	require.True(t, c.Metrics.Global.RemoteWrite[0].SendExemplars)
 }
 
-func TestLoadDynamicConfigurationExpandError(t *testing.T) {
-	err := LoadDynamicConfiguration("", true, nil)
-	assert.Error(t, err)
-	assert.True(t, strings.Contains(err.Error(), "expand var is not supported when using dynamic configuration, use gomplate env instead"))
-}
-
 func TestAgent_OmitEmptyFields(t *testing.T) {
 	var cfg Config
 	yml, err := yaml.Marshal(&cfg)
@@ -473,9 +467,9 @@ agent_management:
     username: "initial_user"
   protocol: "http"
   polling_interval: "1m"
-  remote_config_cache_location: "/etc"
   remote_configuration:
-    namespace: "new_namespace"`
+    namespace: "new_namespace"
+    cache_location: "/etc"`
 
 	remoteCfg := `
 server:
@@ -493,9 +487,9 @@ agent_management:
     username: "new_user"
   protocol: "http"
   polling_interval: "10s"
-  remote_config_cache_location: "/etc"
   remote_configuration:
-    namespace: "new_namespace"`
+    namespace: "new_namespace"
+    cache_location: "/etc"`
 
 	var ic, rc Config
 	err := LoadBytes([]byte(initialCfg), false, &ic)
