@@ -11,6 +11,7 @@ import (
 
 	"github.com/prometheus/client_golang/prometheus"
 	"go.opentelemetry.io/otel/trace"
+	"go.uber.org/atomic"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/component"
@@ -22,8 +23,9 @@ type Controller struct {
 	reg component.Registration
 	log log.Logger
 
-	onRun   sync.Once
-	running chan struct{}
+	onRun    sync.Once
+	running  chan struct{}
+	runError atomic.Error
 
 	innerMut sync.Mutex
 	inner    component.Component
@@ -83,6 +85,9 @@ func (c *Controller) WaitRunning(timeout time.Duration) error {
 	case <-time.After(timeout):
 		return fmt.Errorf("timed out waiting for the controller to start running")
 	case <-c.running:
+		if err := c.runError.Load(); err != nil {
+			return fmt.Errorf("component failed to start: %w", err)
+		}
 		return nil
 	}
 }
@@ -119,11 +124,17 @@ func (c *Controller) Run(ctx context.Context, args component.Arguments) error {
 	}()
 
 	run, err := c.buildComponent(dataPath, args)
+
+	// We close c.running before checking the error, since the component will
+	// never run if we return an error anyway.
+	c.onRun.Do(func() {
+		c.runError.Store(err)
+		close(c.running)
+	})
+
 	if err != nil {
 		return err
 	}
-
-	c.onRun.Do(func() { close(c.running) })
 	return run.Run(ctx)
 }
 
