@@ -49,6 +49,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net"
 	"sync"
 	"time"
 
@@ -110,9 +111,8 @@ type Options struct {
 	// different value for HTTPPathPrefix to prevent components from colliding.
 	HTTPPathPrefix string
 
-	// HTTPListenAddr is the base address that the server is listening on.
-	// The controller does not itself listen here, but some components
-	// need to know this to set the correct targets.
+	// HTTPListenAddr is the base address (host:port) where component APIs are
+	// exposed to other components.
 	HTTPListenAddr string
 
 	// OnExportsChange is called when the exports of the controller change.
@@ -123,6 +123,10 @@ type Options struct {
 
 	// Controller is used to instantiate new module controllers.
 	Controller component.Controller
+
+	// DialFunc is a function to use for components to properly connect to
+	// HTTPListenAddr. If nil, DialFunc defaults to (&net.Dialer{}).DialContext.
+	DialFunc func(ctx context.Context, network, address string) (net.Conn, error)
 }
 
 // Flow is the Flow system.
@@ -160,6 +164,11 @@ func New(o Options) *Flow {
 		}
 	}
 
+	dialFunc := o.DialFunc
+	if dialFunc == nil {
+		dialFunc = (&net.Dialer{}).DialContext
+	}
+
 	var (
 		queue  = controller.NewQueue()
 		sched  = controller.NewScheduler()
@@ -177,6 +186,7 @@ func New(o Options) *Flow {
 			Registerer:      o.Reg,
 			HTTPPathPrefix:  o.HTTPPathPrefix,
 			HTTPListenAddr:  o.HTTPListenAddr,
+			DialFunc:        dialFunc,
 			ControllerID:    o.ControllerID,
 			Controller:      o.Controller,
 		})
