@@ -5,6 +5,7 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/grafana/agent/pkg/river/internal/value"
 	"github.com/grafana/agent/pkg/river/parser"
 	"github.com/grafana/agent/pkg/river/vm"
 	"github.com/stretchr/testify/require"
@@ -19,11 +20,48 @@ func TestVM_Stdlib(t *testing.T) {
 		expect interface{}
 	}{
 		{"env", `env("TEST_VAR")`, string("Hello!")},
-		{"concat", `concat([true, "foo"], [], [false, 1])`, []interface{}{true, "foo", false, 1}},
+		{"concat", `concat([true, "foo"], [], [false, 1])`, []interface{}{true, "foo", false, float64(1)}},
 		{"json_decode object", `json_decode("{\"foo\": \"bar\"}")`, map[string]interface{}{"foo": "bar"}},
 		{"json_decode array", `json_decode("[0, 1, 2]")`, []interface{}{float64(0), float64(1), float64(2)}},
 		{"json_decode nil field", `json_decode("{\"foo\": null}")`, map[string]interface{}{"foo": nil}},
 		{"json_decode nil array element", `json_decode("[0, null]")`, []interface{}{float64(0), nil}},
+	}
+
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			expr, err := parser.ParseExpression(tc.input)
+			require.NoError(t, err)
+
+			eval := vm.New(expr)
+
+			rv := reflect.New(reflect.TypeOf(tc.expect))
+			require.NoError(t, eval.Evaluate(nil, rv.Interface()))
+			require.Equal(t, tc.expect, rv.Elem().Interface())
+		})
+	}
+}
+
+func TestStdlibCoalesce(t *testing.T) {
+	t.Setenv("TEST_VAR2", "Hello!")
+
+	tt := []struct {
+		name   string
+		input  string
+		expect interface{}
+	}{
+		{"coalesce()", `coalesce()`, value.Null},
+		{"coalesce(string)", `coalesce("Hello!")`, string("Hello!")},
+		{"coalesce(string, string)", `coalesce(env("TEST_VAR2"), "World!")`, string("Hello!")},
+		{"(string, string) with fallback", `coalesce(env("NON_DEFINED"), "World!")`, string("World!")},
+		{"coalesce(list, list)", `coalesce([], ["fallback"])`, []string{"fallback"}},
+		{"coalesce(list, list) with fallback", `coalesce(concat(["item"]), ["fallback"])`, []string{"item"}},
+		{"coalesce(int, int, int)", `coalesce(0, 1, 2)`, 1},
+		{"coalesce(bool, int, int)", `coalesce(false, 1, 2)`, 1},
+		{"coalesce(bool, bool)", `coalesce(false, true)`, true},
+		{"coalesce(list, bool)", `coalesce(json_decode("[]"), true)`, true},
+		{"coalesce(object, true) and return true", `coalesce(json_decode("{}"), true)`, true},
+		{"coalesce(object, false) and return false", `coalesce(json_decode("{}"), false)`, false},
+		{"coalesce(list, nil)", `coalesce([],null)`, value.Null},
 	}
 
 	for _, tc := range tt {
